+++ pica-path/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this crate will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [Unreleased]
+
+- #577 Allow optional subfield matcher in a path expression
+
 ## v0.1.0
 
 ### Added